+++ src/main/java/dev/nmarulo/depensaapp/app/productshoppinglist/ProductHasShoppingListRepository.java
@@ -15,19 +15,19 @@ public interface ProductHasShoppingListRepository extends JpaRepository<ProductH
         "SELECT p FROM ProductHasShoppingList p WHERE p.shoppingList.id = :shoppingListId AND p.shoppingList.user = :user AND p.product.id = :productId")
     Optional<ProductHasShoppingList> findByShoppingListIdAndUserAndProductId(Integer shoppingListId,
                                                                              User user,
-                                                                             Integer productId);
+                                                                             Long productId);
     
     @Query(
         "SELECT p FROM ProductHasShoppingList p WHERE p.shoppingList.id = :id AND p.shoppingList.user = :user AND p.product.id IN :productsId")
     List<ProductHasShoppingList> findAllByShoppingListIdAndUserAndProductIdIn(Integer id,
                                                                               User user,
-                                                                              List<Integer> productsId);
+                                                                              List<Long> productsId);
     
     @Query(
         "SELECT p FROM ProductHasShoppingList p WHERE p.shoppingList.id = :id AND p.shoppingList.user = :user AND p.product.id IN :productsId AND p.unitType.id IN :unitTypesId")
     List<ProductHasShoppingList> findAllByShoppingListIdAndUserAndProductIdInAndUnitTypeIdIn(Integer id,
                                                                                              User user,
-                                                                                             List<Integer> productsId,
+                                                                                             List<Long> productsId,
                                                                                              List<Integer> unitTypesId);
     
 }
